+++ src/App.js
@@ -1,22 +1,16 @@
-import React from 'react';
-import {
-  BrowserRouter as Router,
-  Switch,
-  Route,
-  Link
-} from "react-router-dom";
+import React, { useState, useEffect } from "react";
+import { BrowserRouter as Router, Switch, Route, Link } from "react-router-dom";
 
-import './App.css';
+import "./App.css";
 
-
-import MovieSearch from './components/MovieSearch';
-
+import MovieSearch from "./components/MovieSearch";
+import RentalLibrary from "./components/RentalLibrary"
 
 const App = (props) => {
- 
-    return  (
-      <div> 
-       <Router>
+
+  return (
+    <div>
+      <Router>
         <div>
           <nav>
             <ul>
@@ -31,8 +25,8 @@ const App = (props) => {
               </li>
             </ul>
           </nav>
-  
-           <Switch>
+
+          <Switch>
             <Route path="/search">
               <MovieSearch url="http://localhost:3000/" />
             </Route>
@@ -42,38 +36,15 @@ const App = (props) => {
             <Route path="/">
               <Home />
             </Route>
-          </Switch> 
+          </Switch>
         </div>
-      </Router> 
-      </div>
-    );
-    
+      </Router>
+    </div>
+  );
 };
 
-
 function Home() {
-  return (
-  <h1> hello </h1>
-  );
-}
-
-
-
-function RentalLibrary() {
-
- 
-  const [library, setLibrary] = useState([]);
-
-  let newLibrary = [...library]
-  newLibrary.push(movie)
-
-  setLibrary(newLibrary)
-
-  return (
-    
-  );
+  return (<h1> hello </h1>);
 }
 
 export default App;
-
-
